+++ extension/user/frontendLogout.js
@@ -1,0 +1,5 @@
+const UnauthorizedError = require('../models/Errors/UnauthorizedError')
+
+module.exports = async () => {
+  throw new UnauthorizedError()
+}